+++ src/App.js
@@ -12,7 +12,7 @@ import { Lego } from "./pages/hobbies/Lego";
 import { Art } from "./pages/hobbies/Art";
 import { Cosplays } from "./pages/hobbies/Cosplays";
 import { Photography } from "./pages/hobbies/Photography";
-import { Travel } from "./pages/hobbies/Travel";
+import { Travel } from "./pages/hobbies/travel/Travel";
 import { Words } from "./pages/words/Words";
 import { LocateEntrance } from "./pages/amazonian mystery/LocateEntrance";
 import { OpenEntrance } from "./pages/amazonian mystery/OpenEntrance";
